+++ src/views/SignInPage/components/SignInForm/style.css.ts
@@ -1,0 +1,11 @@
+import { style } from '@vanilla-extract/css';
+
+export const inputWrapper = style({
+  display: 'flex',
+  flexDirection: 'column',
+  gap: 50,
+});
+
+export const newPasswordButton = style({
+  borderBottom: '1px solid currentColor',
+});
